refactor: Moved graphics related classes to a specific directory
Updated references and includes etc.
