+++ include/moth_ui/animation/animation_track.h
@@ -3,7 +3,6 @@
 #include "moth_ui/ui_fwd.h"
 
 #include <nlohmann/json_fwd.hpp>
-
 #include <array>
 #include <functional>
