+++ include/moth_ui/animation/clip_controller.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "moth_ui/node.h"
+#include "moth_ui/nodes/node.h"
 #include "moth_ui/animation/animation_clip.h"
 
 namespace moth_ui {
