+++ include/moth_ui/graphics/image_scale_type.h
