+++ include/moth_ui/ifont_factory.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "moth_ui/ifont.h"
+#include "moth_ui/graphics/ifont.h"
 
 #include <filesystem>
 #include <vector>
