+++ include/moth_ui/layout/layout.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "moth_ui/context.h"
-#include "moth_ui/ui_fwd.h"
 #include "moth_ui/layout/layout_entity_group.h"
+#include "moth_ui/ui_fwd.h"
 
 namespace moth_ui {
     class Layout : public LayoutEntityGroup {
