+++ include/moth_ui/layout/layout_entity.h
@@ -1,11 +1,11 @@
 #pragma once
 
-#include "moth_ui/context.h"
-#include "moth_ui/ui_fwd.h"
 #include "moth_ui/animation/animation_track.h"
+#include "moth_ui/context.h"
+#include "moth_ui/graphics/blend_mode.h"
 #include "moth_ui/layout/layout_entity_type.h"
+#include "moth_ui/ui_fwd.h"
 #include "moth_ui/utils/color.h"
-#include "moth_ui/blend_mode.h"
 
 #include <nlohmann/json_fwd.hpp>
