+++ include/moth_ui/layout/layout_entity_ref.h
@@ -1,10 +1,9 @@
 #pragma once
 
-#include "moth_ui/ui_fwd.h"
 #include "moth_ui/layout/layout_entity_group.h"
+#include "moth_ui/ui_fwd.h"
 
 #include <nlohmann/json_fwd.hpp>
-
 #include <filesystem>
 
 namespace moth_ui {
