+++ include/moth_ui/utils/color.h
@@ -1,7 +1,7 @@
 #pragma once
 
+#include "moth_ui/graphics/blend_mode.h"
 #include "moth_ui/utils/vector.h"
-#include "moth_ui/blend_mode.h"
 
 #include <algorithm>
 #include <cmath>
